Fixing a import loop issue
