+++ play/src/front/Chat/Components/Room/RoomTimeline.svelte
@@ -3,7 +3,8 @@
     import { get } from "svelte/store";
     import { ChatRoom } from "../../Connection/ChatConnection";
     import getCloseImg from "../../images/get-close.png";
-    import { selectedChatMessageToReply, selectedRoomStore } from "../../Stores/ChatStore";
+    import { selectedChatMessageToReply } from "../../Stores/ChatStore";
+    import { selectedRoomStore } from "../../Stores/SelectRoomStore";
     import Avatar from "../Avatar.svelte";
     import { matrixSecurity } from "../../Connection/Matrix/MatrixSecurity";
     import { localUserStore } from "../../../Connection/LocalUserStore";
