+++ play/src/front/Chat/Components/RoomList.svelte
@@ -4,7 +4,8 @@
     import { onDestroy, onMount } from "svelte";
     import { gameManager } from "../../Phaser/Game/GameManager";
     import LL from "../../../i18n/i18n-svelte";
-    import { chatSearchBarValue, joignableRoom, navChat, selectedRoomStore } from "../Stores/ChatStore";
+    import { chatSearchBarValue, joignableRoom, navChat } from "../Stores/ChatStore";
+    import { selectedRoomStore } from "../Stores/SelectRoomStore";
     import { ChatRoom } from "../Connection/ChatConnection";
     import { INITIAL_SIDEBAR_WIDTH, loginTokenErrorStore } from "../../Stores/ChatStore";
     import { userIsConnected } from "../../Stores/MenuStore";
