+++ play/src/front/Chat/Connection/Matrix/InteractiveAuthPhase.ts
@@ -1,0 +1,4 @@
+export enum INTERACTIVE_AUTH_PHASE {
+    PRE_AUTH = 1,
+    POST_AUTH,
+}
