+++ play/src/front/Chat/Stores/AlreadyAskForInitCryptoConfigurationStore.ts
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export const alreadyAskForInitCryptoConfiguration = writable(false);
