+++ play/src/front/Chat/Stores/ChatStore.ts
@@ -1,10 +1,10 @@
 import { get, writable } from "svelte/store";
 import { ComponentType } from "svelte";
-import { ChatMessage as NewChatMessage, ChatRoom } from "../Connection/ChatConnection";
+import { ChatMessage as NewChatMessage } from "../Connection/ChatConnection";
 import { chatVisibilityStore } from "../../Stores/ChatStore";
-import { matrixSecurity } from "../Connection/Matrix/MatrixSecurity";
 import { ENABLE_CHAT } from "../../Enum/EnvironmentVariable";
 import { gameManager } from "../../Phaser/Game/GameManager";
+import { selectedRoomStore } from "./SelectRoomStore";
 
 type NavChatTab =
     | {
@@ -48,34 +48,6 @@ export const navChat = createNavChatStore();
 export const shownRoomListStore = writable<string>("");
 export const chatSearchBarValue = writable<string>("");
 
-const createSelectedRoomStore = () => {
-    const { subscribe, update } = writable<ChatRoom | undefined>(undefined);
-
-    const customSet = (value: ChatRoom | undefined) => {
-        update((currentValue) => {
-            if (
-                currentValue !== value &&
-                value &&
-                get(value.isEncrypted) &&
-                !get(alreadyAskForInitCryptoConfiguration)
-            ) {
-                matrixSecurity.openChooseDeviceVerificationMethodModal().catch((error) => {
-                    console.error(error);
-                });
-            }
-
-            return value;
-        });
-    };
-
-    return {
-        subscribe,
-        set: customSet,
-    };
-};
-
-export const selectedRoomStore = createSelectedRoomStore();
-
 export const selectedChatMessageToReply = writable<NewChatMessage | null>(null);
 
 export const selectedChatMessageToEdit = writable<NewChatMessage | null>(null);
@@ -85,8 +57,6 @@ export const joignableRoom = writable<{ id: string; name: string | undefined }[]
 export const isAChatRoomIsVisible = () => {
     return get(selectedRoomStore) && get(navChat).key === "chat" && get(chatVisibilityStore);
 };
-
-export const alreadyAskForInitCryptoConfiguration = writable(false);
 
 export const isChatIdSentToPusher = writable(false);
