--- play/src/front/Notification/index.ts
@@ -1,6 +1,0 @@
-import { MessageNotification } from "./MessageNotification";
-import { BubbleNotification } from "./BubbleNotification";
-import { NotificationWA } from "./Notification";
-import { notificationManager } from "./NotificationManager";
-export { MessageNotification, BubbleNotification as BasicNotification, notificationManager };
-export type { NotificationWA };
