+++ play/src/front/WebRtc/SimplePeer.ts
@@ -15,7 +15,8 @@ import { analyticsClient } from "../Administration/AnalyticsClient";
 import { nbSoundPlayedInBubbleStore } from "../Stores/ApparentMediaContraintStore";
 import { SpaceFilterInterface } from "../Space/SpaceFilter/SpaceFilter";
 import { RemotePlayersRepository } from "../Phaser/Game/RemotePlayersRepository";
-import { BasicNotification, notificationManager } from "../Notification";
+import { BubbleNotification as BasicNotification } from "../Notification/BubbleNotification";
+import { notificationManager } from "../Notification/NotificationManager";
 import LL from "../../i18n/i18n-svelte";
 import { mediaManager } from "./MediaManager";
 import { ScreenSharingPeer } from "./ScreenSharingPeer";
